+++ sentry-android-replay/src/main/java/io/sentry/android/replay/util/Views.kt
@@ -13,6 +13,8 @@ import android.graphics.drawable.VectorDrawable
 import android.os.Build.VERSION
 import android.os.Build.VERSION_CODES
 import android.text.Layout
+import android.text.Spanned
+import android.text.style.ForegroundColorSpan
 import android.view.View
 import android.widget.TextView
 import java.lang.NullPointerException
@@ -101,3 +103,27 @@ internal val TextView.totalPaddingTopSafe: Int
     } catch (e: NullPointerException) {
         extendedPaddingTop
     }
+
+internal val TextView.dominantTextColor: Int get() {
+    if (text !is Spanned) return currentTextColor
+
+    val spans = (text as Spanned).getSpans(0, text.length, ForegroundColorSpan::class.java)
+
+    // determine the dominant color by the span with the longest range
+    var longestSpan = Int.MIN_VALUE
+    var dominantColor: Int? = null
+    for (span in spans) {
+        val spanStart = (text as Spanned).getSpanStart(span)
+        val spanEnd = (text as Spanned).getSpanEnd(span)
+        if (spanStart == -1 || spanEnd == -1) {
+            // the span is not attached
+            continue
+        }
+        val spanLength = spanEnd - spanStart
+        if (spanLength > longestSpan) {
+            longestSpan = spanLength
+            dominantColor = span.foregroundColor
+        }
+    }
+    return dominantColor ?: currentTextColor
+}
